third_party/WebKit/LayoutTests/fast/loader/navigation-scheduler-user-gesture-expected.txt

Check that the NavigationScheduler correctly tracks user gestures

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.open('about:blank') is undefined.
 click me
